Create freshdeals-blueberries.html
Added breadcrumbs to this site. By doing so, users will have a better feeling for what the site provides and the optional attributes of the product they are purchasing.

Focused only on the layout of this page.  Implemented the breadcrumbs via HTML and CSS. The links to the breadcrumbs will not function.